typescript-react-redux-combine-state-reducers-demo

+++ src/components/hello.tsx
@@ -1,39 +1,55 @@
 import React, {Dispatch} from 'react';
 import {connect} from 'react-redux';
-import {State} from '../store/State';
-import {DeleteFruitAction} from '../store/actions/actionTypes';
-import {deleteFruitAction} from '../store/actions';
+import {ApplicationActions, ApplicationState} from '../store/application';
+import {deleteFruitAction} from '../store/apiState/actions/apiStateActions';
+import {toggleDeleteButtons} from '../store/formOnlyState/actions/actionTypes';
 
 type StateProps = {
+  showDeleteButtons?: boolean,
   fruits: string[]
 }
 
 type DispatchProps = {
-  deleteFruit: (fruitName: string) => void
+  deleteFruit: (fruitName: string) => void,
+  toggleDeleteButtons: () => void
 }
 
 type Props = StateProps & DispatchProps;
 
-function Hello({fruits, deleteFruit}: Props) {
+function Hello(
+  {
+    showDeleteButtons,
+    fruits,
+    deleteFruit,
+    toggleDeleteButtons
+  }: Props
+) {
   return <div>
     <h1>Hello, I like:</h1>
     <ul>
       {
         fruits.map(name => <li key={name}>
           {name}
-          <button title="delete" onClick={() => deleteFruit(name)}>x</button>
+          {
+            showDeleteButtons && <button title="delete" onClick={() => deleteFruit(name)}>x</button>
+          }
         </li>)
       }
     </ul>
+    <div>
+      <button onClick={toggleDeleteButtons}>Show/Hide Delete Buttons</button>
+    </div>
   </div>
 }
 
-const mapStateToProps = (state: State): StateProps => ({
-  fruits: state.fruits
+const mapStateToProps = (state: ApplicationState): StateProps => ({
+  fruits: state.apiState.fruits,
+  showDeleteButtons: state.formOnlyState.showDeleteButtons
 })
 
-const mapDispatchToProps = (dispatch: Dispatch<DeleteFruitAction>): DispatchProps => ({
-  deleteFruit: (name) => dispatch(deleteFruitAction(name))
+const mapDispatchToProps = (dispatch: Dispatch<ApplicationActions>): DispatchProps => ({
+  deleteFruit: (name) => dispatch(deleteFruitAction(name)),
+  toggleDeleteButtons: () => dispatch(toggleDeleteButtons()),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(Hello);
